FCN-123: consolidated types into single file
